# src/Domivice.Users.Web/Models/DayOff.cs
/*
 * Users API
 *
 * The users API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using System.Text;
using Newtonsoft.Json;

namespace Domivice.Users.Web.Models;

/// <summary>
///     Carries information about a day off
/// </summary>
[DataContract]
public class DayOff : IEquatable<DayOff>
{
    /// <summary>
    ///     The string guid representation of the day off id
    /// </summary>
    /// <value>The string guid representation of the day off id</value>
    [Required]
    [DataMember(Name = "id", EmitDefaultValue = false)]
    public string Id { get; set; }

    /// <summary>
    ///     The day off user id
    /// </summary>
    /// <value>The day off user id</value>
    [Required]
    [DataMember(Name = "userId", EmitDefaultValue = false)]
    public string UserId { get; set; }

    /// <summary>
    ///     The day off start date
    /// </summary>
    /// <value>The day off start date</value>
    [Required]
    [DataMember(Name = "startDate", EmitDefaultValue = false)]
    public DateTime StartDate { get; set; }

    /// <summary>
    ///     The day off end date
    /// </summary>
    /// <value>The day off end date</value>
    [Required]
    [DataMember(Name = "endDate", EmitDefaultValue = false)]
    public DateTime EndDate { get; set; }

    /// <summary>
    ///     Returns true if DayOff instances are equal
    /// </summary>
    /// <param name="other">Instance of DayOff to be compared</param>
    /// <returns>Boolean</returns>
    public bool Equals(DayOff other)
    {
        if (other is null) return false;
        if (ReferenceEquals(this, other)) return true;

        return
            (
                Id == other.Id ||
                (Id != null &&
                 Id.Equals(other.Id))
            ) &&
            (
                UserId == other.UserId ||
                (UserId != null &&
                 UserId.Equals(other.UserId))
            ) &&
            (
                StartDate == other.StartDate ||
                (StartDate != null &&
                 StartDate.Equals(other.StartDate))
            ) &&
            (
                EndDate == other.EndDate ||
                (EndDate != null &&
                 EndDate.Equals(other.EndDate))
            );
    }

    /// <summary>
    ///     Returns the string presentation of the object
    /// </summary>
    /// <returns>String presentation of the object</returns>
    public override string ToString()
    {
        var sb = new StringBuilder();
        sb.Append("class DayOff {\n");
        sb.Append("  Id: ").Append(Id).Append("\n");
        sb.Append("  UserId: ").Append(UserId).Append("\n");
        sb.Append("  StartDate: ").Append(StartDate).Append("\n");
        sb.Append("  EndDate: ").Append(EndDate).Append("\n");
        sb.Append("}\n");
        return sb.ToString();
    }

    /// <summary>
    ///     Returns the JSON string presentation of the object
    /// </summary>
    /// <returns>JSON string presentation of the object</returns>
    public string ToJson()
    {
        return JsonConvert.SerializeObject(this, Formatting.Indented);
    }

    /// <summary>
    ///     Returns true if objects are equal
    /// </summary>
    /// <param name="obj">Object to be compared</param>
    /// <returns>Boolean</returns>
    public override bool Equals(object obj)
    {
        if (obj is null) return false;
        if (ReferenceEquals(this, obj)) return true;
        return obj.GetType() == GetType() && Equals((DayOff)obj);
    }

    /// <summary>
    ///     Gets the hash code
    /// </summary>
    /// <returns>Hash code</returns>
    public override int GetHashCode()
    {
        unchecked // Overflow is fine, just wrap
        {
            var hashCode = 41;
            // Suitable nullity checks etc, of course :)
            if (Id != null)
                hashCode = hashCode * 59 + Id.GetHashCode();
            if (UserId != null)
                hashCode = hashCode * 59 + UserId.GetHashCode();
            if (StartDate != null)
                hashCode = hashCode * 59 + StartDate.GetHashCode();
            if (EndDate != null)
                hashCode = hashCode * 59 + EndDate.GetHashCode();
            return hashCode;
        }
    }

    #region Operators

#pragma warning disable 1591

    public static bool operator ==(DayOff left, DayOff right)
    {
        return Equals(left, right);
    }

    public static bool operator !=(DayOff left, DayOff right)
    {
        return !Equals(left, right);
    }

#pragma warning restore 1591

    #endregion Operators
}