# src/Domivice.Users.Web/wwwroot/openapi-original.json
{
    "openapi": "3.0.2",
    "info": {
        "contact": {
            "name": "Domivice Development Team",
            "url": "https://development.domivice.com"
        },
        "description": "The users API",
        "title": "Users API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
...
    "paths": {
...
        "/v1/users/{userId}/days-off": {
            "get": {
                "parameters": [
                    {
                        "description": "The string guid representing the user id",
                        "explode": false,
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DaysOffList"
                                }
                            }
                        },
                        "description": "Response when a list of 0 or more days off is returned"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "User is not authenticated"
                    }
                },
                "summary": "List Days Off",
                "tags": ["Days Off"]
            },
            "post": {
                "parameters": [
                    {
                        "description": "The string guid representing the user id",
                        "explode": false,
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DayOffCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DayOff"
                                }
                            }
                        },
                        "description": "Response when a day off is created successfully"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "There was an error with the request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "User is not authenticated"
                    }
                },
                "summary": "Create Day Off",
                "tags": ["Days Off"]
            }
        },
        "/v1/users/{userId}/days-off/{dayOffId}": {
            "delete": {
                "parameters": [
                    {
                        "description": "The string guid representing the user id",
                        "explode": false,
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }, {
                        "description": "The day off id",
                        "explode": false,
                        "in": "path",
                        "name": "dayOffId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Response when the day off we deleted."
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "User is not authenticated"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "User is not allowed to perform action"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "The resource was not found"
                    }
                },
                "summary": "Delete Day Off",
                "tags": ["Days Off"]
            },
            "get": {
                "parameters": [
                    {
                        "description": "The string guid representing the user id",
                        "explode": false,
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }, {
                        "description": "The day off id",
                        "explode": false,
                        "in": "path",
                        "name": "dayOffId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DayOff"
                                }
                            }
                        },
                        "description": "The response when a day off object is returned"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "User is not authenticated"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "The resource was not found"
                    }
                },
                "summary": "Get Day Off",
                "tags": ["Days Off"]
            },
            "patch": {
                "parameters": [
                    {
                        "description": "The string guid representing the user id",
                        "explode": false,
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }, {
                        "description": "The day off id",
                        "explode": false,
                        "in": "path",
                        "name": "dayOffId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DayOffUpdate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DayOff"
                                }
                            }
                        },
                        "description": "Response when the day off was updated successfully."
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "There was an error with the request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "User is not authenticated"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "The resource was not found"
                    }
                },
                "summary": "Patch Day Off",
                "tags": ["Days Off"]
            }
        }
    },
    "components": {
...
        "schemas": {
            "Error": {
                "description": "",
                "properties": {
                    "detail": {
                        "description": "A human-readable explanation of the error",
                        "type": "string"
                    },
                    "instance": {
                        "description": "A URI that identifies the specific occurrence of the error",
                        "type": "string"
                    },
                    "status": {
                        "description": "The HTTP response code",
                        "type": "string"
                    },
                    "title": {
                        "description": "A brief, human-readable message about the error",
                        "type": "string"
                    },
                    "type": {
                        "description": "A URI identifier that categorizes the error",
                        "type": "string"
                    },
                    "errors": {
                        "description": "The validation errors",
                        "items": {
                            "$ref": "#/components/schemas/KeyValuePair"
                        },
                        "type": "array"
                    },
                    "extensions": {
                        "description": "The error extensions",
                        "items": {
                            "$ref": "#/components/schemas/KeyValuePair"
                        },
                        "type": "array"
                    }
                },
                "required": ["detail", "instance", "title", "type"],
                "type": "object"
            },
            "KeyValuePair": {
                "description": "Carries key values pair",
                "type": "object"
            },
...
            "DayOff": {
                "description": "Carries information about a day off",
                "example": {
                    "endDate": "2000-01-23T04:56:07.000+00:00",
                    "id": "id",
                    "userId": "userId",
                    "startDate": "2000-01-23T04:56:07.000+00:00"
                },
                "properties": {
                    "id": {
                        "description": "The string guid representation of the day off id",
                        "type": "string"
                    },
                    "userId": {
                        "description": "The day off user id",
                        "type": "string"
                    },
                    "startDate": {
                        "description": "The day off start date",
                        "format": "date-time",
                        "type": "string"
                    },
                    "endDate": {
                        "description": "The day off end date",
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": ["endDate", "id", "startDate", "userId"],
                "type": "object"
            },
            "DayOffCreate": {
                "description": "Carries information about a day off",
                "example": {
                    "endDate": "2000-01-23T04:56:07.000+00:00",
                    "startDate": "2000-01-23T04:56:07.000+00:00"
                },
                "properties": {
                    "startDate": {
                        "description": "The day off start date",
                        "format": "date-time",
                        "type": "string"
                    },
                    "endDate": {
                        "description": "The day off end date",
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": ["endDate", "startDate"],
                "type": "object"
            },
            "DayOffUpdate": {
                "description": "Carries information about a day off",
                "example": {
                    "endDate": "2000-01-23T04:56:07.000+00:00",
                    "startDate": "2000-01-23T04:56:07.000+00:00"
                },
                "properties": {
                    "startDate": {
                        "description": "The day off start date",
                        "format": "date-time",
                        "type": "string"
                    },
                    "endDate": {
                        "description": "The day off end date",
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "DaysOffList": {
                "description": "Carries a paginated list of users",
                "example": {
                    "totalItemsCount": 6,
                    "pageCount": 0,
                    "previousPage": "previousPage",
                    "data": [
                        {
                            "endDate": "2000-01-23T04:56:07.000+00:00",
                            "id": "id",
                            "userId": "userId",
                            "startDate": "2000-01-23T04:56:07.000+00:00"
                        }, {
                            "endDate": "2000-01-23T04:56:07.000+00:00",
                            "id": "id",
                            "userId": "userId",
                            "startDate": "2000-01-23T04:56:07.000+00:00"
                        }
                    ],
                    "nextPage": "nextPage"
                },
                "properties": {
                    "data": {
                        "description": "The days off list",
                        "items": {
                            "$ref": "#/components/schemas/DayOff"
                        },
                        "type": "array"
                    },
                    "nextPage": {
                        "description": "The URL to access the next page",
                        "type": "string"
                    },
                    "pageCount": {
                        "description": "The page count",
                        "type": "integer"
                    },
                    "previousPage": {
                        "description": "The URL to access the previous page",
                        "type": "string"
                    },
                    "totalItemsCount": {
                        "description": "The total items count",
                        "type": "integer"
                    }
                },
                "required": ["data", "pageCount", "totalItemsCount"],
                "type": "object"
            },
...
        }
    }
}